# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tidarr API",
    "description": "Tidarr REST API for managing Tidal music downloads, queue management, synchronization, and Lidarr integration.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8484",
      "description": "Local Tidarr instance"
    }
  ],
  "security": [
    {
      "ApiKeyHeader": []
    },
    {
      "ApiKeyQuery": []
    },
    {
      "BearerToken": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "API key passed via X-Api-Key header (recommended for automation)"
      },
      "ApiKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "apikey",
        "description": "API key passed via query parameter"
      },
      "BearerToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from POST /api/auth"
      }
    },
    "schemas": {
...
      "AuthErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "SignedUrlResponse": {
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "url": {
            "type": "string",
            "description": "Signed URL path with expiration and signature parameters"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/auth/oidc/login": {
      "get": {
        "operationId": "oidcLogin",
        "summary": "Initiate OIDC login",
        "description": "Redirects to the configured OIDC provider for authentication. Requires OIDC_ISSUER, OIDC_CLIENT_ID, OIDC_CLIENT_SECRET, and OIDC_REDIRECT_URI environment variables.",
        "security": [],
        "responses": {
          "302": {
            "description": "Redirect to OIDC provider"
          },
          "500": {
            "description": "OIDC configuration incomplete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/oidc/callback": {
      "get": {
        "operationId": "oidcCallback",
        "summary": "OIDC callback handler",
        "description": "Handles the OIDC callback, exchanges the authorization code for tokens, and redirects to the frontend with a JWT token.",
        "security": [],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Authorization code from OIDC provider"
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "State parameter for CSRF protection"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to frontend with JWT token"
          },
          "400": {
            "description": "Missing code or state parameter"
          },
          "403": {
            "description": "Invalid state parameter"
          }
        }
      }
    },
    "/api/stream/sign/{id}": {
      "get": {
        "operationId": "signStreamUrl",
        "summary": "Generate signed streaming URL",
        "description": "Generates a time-limited (5 minutes) signed URL for audio streaming of a Tidal track. This endpoint does NOT require authentication.",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Tidal track ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Signed URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedUrlResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing ID"
          }
        }
      }
    },
    "/api/stream/play/{id}": {
      "get": {
        "operationId": "playStream",
        "summary": "Play audio stream",
        "description": "Streams audio content with signature validation. Supports HTTP range requests for seeking. Use the signed URL from /api/stream/sign/:id. This endpoint does NOT require authentication (uses URL signature instead).",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Tidal track ID"
          },
          {
            "name": "exp",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Expiration timestamp"
          },
          {
            "name": "sig",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "URL signature"
          }
        ],
        "responses": {
          "200": {
            "description": "Audio stream",
            "content": {
              "audio/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "No Tidal token available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Missing signature, expired URL, or invalid signature",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "No playable quality available from Tidal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}